backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
manim==0.18.0
python-dotenv==1.0.0
bcrypt==4.0.1
alembic==1.13.0
email-validator==2.1.0
